requirements.txt
django==2.1
django-autoslug-iplweb==1.9.4
pillow==5.2.0
django-bootstrap4==0.0.7
djangorestframework==3.8.2
